Reorg to allow multiple versions of BenchmarkReport schema
