+++ analysis/benchmark_report.py
@@ -1,0 +1,1 @@
+../workload/report/benchmark_report.py
