+++ config_explorer/src/config_explorer/benchmark_report_v0_2.py
@@ -1,0 +1,1 @@
+../../../workload/report/benchmark_report_v0_2.py
