+++ config_explorer/src/config_explorer/explorer.py
@@ -51,7 +51,12 @@ import yaml
 # llm-d-benchmark repository is refactored into full Python.
 try:
     import convert
-    import schema
+    from benchmark_report_v0_1 import (
+        BenchmarkReport,
+        HostType,
+        Units,
+        WorkloadGenerator,
+    )
     from constants import (
         BOUND_PREFIX_LEN,
         COLUMN_BOUND_STR,
@@ -59,7 +64,12 @@ try:
     )
 except ImportError:
     from config_explorer import convert
-    from config_explorer import schema
+    from config_explorer.benchmark_report_v0_1 import (
+        BenchmarkReport,
+        HostType,
+        Units,
+        WorkloadGenerator,
+    )
     from config_explorer.constants import (
         BOUND_PREFIX_LEN,
         COLUMN_BOUND_STR,
@@ -883,7 +893,7 @@ def make_benchmark_runs_df() -> pd.DataFrame:
 
 
 def _get_replicas_and_parallelism(
-        report: schema.BenchmarkReport) -> dict[str, int | None]:
+        report: BenchmarkReport) -> dict[str, int | None]:
     """Get the number of replicas and parallelisms.
 
     Args:
@@ -917,9 +927,9 @@ def _get_replicas_and_parallelism(
         # Host details are not available
         return rp
 
-    rp['replicas'] = report.scenario.host.type.count(schema.HostType.REPLICA)
-    rp['p_replicas'] = report.scenario.host.type.count(schema.HostType.PREFILL)
-    rp['d_replicas'] = report.scenario.host.type.count(schema.HostType.DECODE)
+    rp['replicas'] = report.scenario.host.type.count(HostType.REPLICA)
+    rp['p_replicas'] = report.scenario.host.type.count(HostType.PREFILL)
+    rp['d_replicas'] = report.scenario.host.type.count(HostType.DECODE)
     if rp['replicas'] == 0:
         rp['replicas'] = None
     if rp['p_replicas'] == 0:
@@ -938,12 +948,12 @@ def _get_replicas_and_parallelism(
     # We have a P/D setup
     rp['is_pd'] = True
     for ii, accel in enumerate(report.scenario.host.accelerator):
-        if report.scenario.host.type[ii] is schema.HostType.PREFILL and rp['p_tp'] is None:
+        if report.scenario.host.type[ii] is HostType.PREFILL and rp['p_tp'] is None:
             rp['p_tp'] = accel.parallelism.tp
             rp['p_dp'] = accel.parallelism.dp
             rp['p_pp'] = accel.parallelism.pp
             rp['p_ep'] = accel.parallelism.ep
-        if report.scenario.host.type[ii] is schema.HostType.DECODE and rp['d_tp'] is None:
+        if report.scenario.host.type[ii] is HostType.DECODE and rp['d_tp'] is None:
             rp['d_tp'] = accel.parallelism.tp
             rp['d_dp'] = accel.parallelism.dp
             rp['d_pp'] = accel.parallelism.pp
@@ -1039,7 +1049,7 @@ def add_benchmark_report_to_df(
     prompts_per_group = None    # Common prefixes within a group
     target_osl = None
     args = report.scenario.load.args
-    if report.scenario.load.name == schema.WorkloadGenerator.INFERENCE_PERF:
+    if report.scenario.load.name == WorkloadGenerator.INFERENCE_PERF:
         # Workload generator stage
         # If stage metadata is not present in benchmark report, we cannot know
         # which Inference Perf result this data came from.
@@ -1061,9 +1071,9 @@ def add_benchmark_report_to_df(
             get_nested(
                 args, [
                     'data', 'shared_prefix', 'output_len'], -1))
-    elif report.scenario.load.name == schema.WorkloadGenerator.VLLM_BENCHMARK:
+    elif report.scenario.load.name == WorkloadGenerator.VLLM_BENCHMARK:
         concurrency = args.get('max_concurrency')
-    elif report.scenario.load.name == schema.WorkloadGenerator.GUIDELLM:
+    elif report.scenario.load.name == WorkloadGenerator.GUIDELLM:
         # Workload generator stage
         # If stage metadata is missing, this benchmark report is from an older
         # version of convert.py that only took stage 0 results.
@@ -1084,10 +1094,10 @@ def add_benchmark_report_to_df(
             target_osl = data.get('output_tokens')
 
     # Multipliers to ensure values are in ms
-    ttft_mult = 1000 if report.metrics.latency.time_to_first_token.units == schema.Units.S else 1
-    tpot_mult = 1000 if report.metrics.latency.time_per_output_token.units == schema.Units.S_PER_TOKEN else 1
-    itl_mult = 1000 if report.metrics.latency.inter_token_latency.units == schema.Units.S_PER_TOKEN else 1
-    e2el_mult = 1000 if report.metrics.latency.request_latency.units == schema.Units.S else 1
+    ttft_mult = 1000 if report.metrics.latency.time_to_first_token.units == Units.S else 1
+    tpot_mult = 1000 if report.metrics.latency.time_per_output_token.units == Units.S_PER_TOKEN else 1
+    itl_mult = 1000 if report.metrics.latency.inter_token_latency.units == Units.S_PER_TOKEN else 1
+    e2el_mult = 1000 if report.metrics.latency.request_latency.units == Units.S else 1
 
     # Calculated metrics
     thpt_per_gpu = None
